# app/BDNR_FLUSK/static/api_specs.yaml
openapi: "3.0.0"
info:
  version: "1.0.0"
  title: "Hydroculus API"
  description: "API untuk mengelola data sumber air, jenisnya, upaya peningkatan, dan pengguna."
paths:
  /api/register:
    post:
      summary: "Registrasi pengguna baru"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
      responses:
        '201':
          description: "Registrasi berhasil"
        '400':
          description: "Username sudah digunakan"
  /api/login:
    post:
      summary: "Login pengguna"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: "Login berhasil"
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
        '401':
          description: "Username atau password salah"
        '400':
          description: "Permintaan tidak valid"
...
  /api/upaya-create:
    post:
      summary: "Menambahkan upaya baru"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                nama_upaya:
                  type: string
      responses:
        '201':
          description: "Upaya berhasil ditambahkan"
  /api/upaya-update/{id}:
    put:
      summary: "Memperbarui upaya berdasarkan ID"
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                nama_upaya:
                  type: string
      responses:
        '200':
          description: "Upaya berhasil diperbarui"
  /api/upaya-delete/{id}:
    delete:
      summary: "Menghapus upaya berdasarkan ID"
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: "Upaya berhasil dihapus"
  /api/sumber_air:
    get:
      summary: "Mendapatkan semua sumber air"
      responses:
        '200':
          description: "Daftar sumber air"
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    _id:
                      type: string
                    nama_sumber_air:
                      type: string
                    id_jenis_sumber_air:
                      type: string
                    upaya_peningkatan:
                      type: array
                      items:
                        type: string
        '500':
          description: "Kesalahan server"
  /api/sumber_air_lookup_filter:
    get:
      summary: "Mencari sumber air berdasarkan keyword dan pagination"
      parameters:
        - name: keyword
          in: query
          required: false
          schema:
            type: string
        - name: page
          in: query
          required: false
          schema:
            type: integer
        - name: limit
          in: query
          required: false
          schema:
            type: integer
      responses:
        '200':
          description: "Daftar sumber air dengan pagination"
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                  total:
                    type: integer
                  page:
                    type: integer
                  limit:
                    type: integer
        '500':
          description: "Kesalahan server"
  /api/sumber_air_lookup:
    get:
      summary: "Mendapatkan semua sumber air dengan lookup"
      responses:
        '200':
          description: "Daftar sumber air dengan lookup"
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    _id:
                      type: string
                    nama_sumber_air:
                      type: string
        '500':
          description: "Kesalahan server"
...
  /api/sumber_air_by_upayas:
    post:
      summary: "Mendapatkan sumber air berdasarkan ID upaya"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id_upayas:
                  type: array
                  items:
                    type: string
      responses:
        '200':
          description: "Daftar sumber air"
          content:
            application/json:
              schema:
                type: object
        '400':
          description: "Permintaan tidak valid"
        '500':
          description: "Kesalahan server"
  /api/sumber_air_create:
    post:
      summary: Add a new sumber air
      description: This endpoint adds a new sumber air to the database.
      operationId: addWater
      tags:
        - Sumber Air
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                nama_sumber_air:
                  type: string
                  description: The name of the water source.
                  example: "Waduk Jati Luhur"
                kondisi_sumber_air:
                  type: string
                  description: The condition of the water source.
                  example: "Rusak Sedang"
                suhu:
                  type: number
                  description: The temperature of the water source.
                  example: 26
                warna:
                  type: string
                  description: The color of the water.
                  example: "Keruh"
                ph:
                  type: number
                  description: The pH level of the water.
                  example: 7.6
                kelayakan:
                  type: string
                  description: The feasibility of the water source.
                  example: "Tidak"
                id_jenis_sumber_air:
                  type: string
                  description: The type of the water source.
                  example: "6766da641f686546f26c201a"
                id_kabupaten:
                  type: string
                  description: The ID of the associated regency.
                  example: "3214"
                foto_sumber_air:
                  type: string
                  description: The filename of the water source photo.
                  example: "foto_waduk_jatiluhur.jpg"
                upaya_peningkatan:
                  type: array
                  items:
                    type: string
                  description: A list of improvement efforts related to the water source.
                  example: ["Pengelolaan Sumber Daya Alam", "Peningkatan Infrastruktur"]
      responses:
        '201':
          description: Sumber air created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Sumber air added successfully"
                  id:
                    type: string
                    description: The ID of the created sumber air
                    example: "6766d9341f686546f26c1fd7"
        '400':
          description: Bad request - Missing required fields
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Missing required fields"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Error message from server"
...
  /api/sumber_air_delete/{id}:
    delete:
      summary: Delete a sumber air
      description: This endpoint deletes a sumber air from the database.
      operationId: deleteWater
      tags:
        - Sumber Air
      parameters:
        - name: id
          in: path
          required: true
          description: The ID of the sumber air to delete.
          schema:
            type: string
            example: "6766d9341f686546f26c1fd7"
      responses:
        '200':
          description: Sumber air deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Sumber air deleted successfully"
                  id:
                    type: string
                    description: The ID of the deleted sumber air
                    example: "6766d9341f686546f26c1fd7"
        '400':
          description: Bad request - Invalid ID format
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Invalid ID format"
        '404':
          description: Sumber air not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Sumber air not found"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Error message from server"
  /protected:
    get:
      summary: "Route yang dilindungi"
      responses:
        '200':
          description: "Akses diterima"
        '401':
          description: "Akses ditolak"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT